requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
supabase==2.5.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
httpx==0.27.0
PyJWT==2.8.0
bcrypt==4.1.3
